move all functions to a sqlquery folder
