+++ service/history/workflow/matcher/evaluators_test.go
@@ -29,7 +29,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	enumspb "go.temporal.io/api/enums/v1"
 	persistencespb "go.temporal.io/server/api/persistence/v1"
-	"go.temporal.io/server/common/util"
+	"go.temporal.io/server/common/sqlquery"
 	"go.temporal.io/server/service/history/workflow"
 	"go.uber.org/mock/gomock"
 	"google.golang.org/protobuf/types/known/timestamppb"
@@ -225,7 +225,7 @@ func TestBasicMutableStateMatchEvaluator(t *testing.T) {
 		},
 	}
 
-	startTime, err := util.ConvertToTime(fmt.Sprintf("'%s'", startTimeStr))
+	startTime, err := sqlquery.ConvertToTime(fmt.Sprintf("'%s'", startTimeStr))
 	assert.NoError(t, err)
 
 	ws := &persistencespb.WorkflowExecutionState{
@@ -307,7 +307,7 @@ func TestAdvancedMutableStateMatchEvaluator(t *testing.T) {
 		},
 	}
 
-	startTime, err := util.ConvertToTime(fmt.Sprintf("'%s'", startTimeStr))
+	startTime, err := sqlquery.ConvertToTime(fmt.Sprintf("'%s'", startTimeStr))
 	assert.NoError(t, err)
 
 	ws := &persistencespb.WorkflowExecutionState{
